feat: use encrypted JWTs for auth
Nativescript doesn't support cookies very well, which made session handling difficult.
Instead, use JWTs, which can be passed with the websocket as a separate protocol
It's not perfect, but it's progress.
Websockets do have a bug, related to consuming messages/reusing the upgrader that manifests when creating a second game on the server.