feat: limit cards drawn per round
Add some minimal server validation, and the concept of game rounds with a history

for speed, and compromise, it's going to make sense to let the client do what it wants, within limits for simple checks like this.

We now have the ability to restrict actions of the user, based on the history of previous actions.

This doesn't extend to user specific actions, although that wouldn't be too difficult.