feat: connect resources to the deck
on producing a good, a card is taken from the deck and added to a reserved cards space in the game state.
Because users don't know what's in the deck, nor do they know what card a resource represents, we can just store it and pick from the queue as we like.

Tests needed updating to accommodate
Also, produce good is now a server response, although when it comes to defining the separation between server/client, this will likely change with the others.