feat: connect the market to the deck, and limit according to suns
This uses actual cards from the deck/discard when populating the market
Also introduces the limit of suns for the cards to be 3.
As a result, more test cards needed to be added.
The CLI is using the name as an ID in some cases, so it's likely this has a bug in a corner case somewhere.
But names won't always be used for IDs, so ignoring for now.