feat: allow placeholder input resources where a building can choose any base input
i regret not handling placeholders as a dedicated item, instead of fudging it in as another type of resource.
Hence, quite a few tests to try mitigate the difficult and tangled logic in the production phase.
Also, it's quite easy to send the market into an infinite loop at the moment due to the small amount of cards.