feat: allow discarding cards at beginning of round
repurpose the discard action. It was originally used for gathering resources for production, but that architecture changed.
Now, multi, unlimited cards can be selected and discarded and drawn to replace them.
We're still using the simple discard/deck model - could do with abstratcing this to a function to shuffle the cards.