feat: add investors as free assignees at the end of the game
couple of ways to achieve this, but this seemed the easiest to start.
I noticed a small bug in producing goods, and left a comment.
Once moving to Vue, it's likely this complicated game loop can be simplified, and the bug will go away, so I'm not addressing it now.
Also adjusted to debug output so it's easier to track the change to the overall game state