refactor: move local-server functions into the client
This way, they're easier to test, and they will fit the real client architecture better.
A step toward defining the server/client boundaries.
We should notify the server on updates, but not update the state based on the response.
In the client, we should listen for a failed/invalid message, and restore the state to the point provided by the message. But this isn't easy to replicate in the CLI, so will probably skip this in the CLI, and just trust the optimistic updates