fix: bug when using lodash difference
difference was uniqifying the results, meaning picking a resource for one value would remove all instances of that resource.

The alternative to this was using instances of resources, rather than importing the same instance across each file. This would need a createResource function(), and I thought about doing that, but it needs several changes throughout the app and overcomplicates things, as most of the schema would the update to using a resourceType instead of a Resource. Just have to be wary of the mutation bugs that could arise as a result of using the same resources imported.