fix: testing conflicts with nativescript
use jest globals to explicitly import the functions
The types loading didn't seem to work, and nativescript doesn't look like it offers jest out of the box, so there's not a lot of advice.
There is a separate plugin just for jest and vue testing
So, we'll revisit this when setting up vue-test-utils.

Ideally, these spec files shouldn't even be in the build.

As part of this explicit importing of the globals, there was another native node package which conflicted with nativescript ("module"), which needed to be added to the shim - https://github.com/azu/rc-config-loader/issues/2