refactor: improve types and simplify available actions
there are now separate enums for server action, server response, and player action.
These are stored as records, which means all enums must have an associated function with a shared signature.
This will help when separating the gameState and server state.
The available actions needed to be converted to normal enums as  a result of this, as the handler wasn't accessible in this way