feat: add unassignment payment
uses a mock employee with a cost
also quickly specifies a cost extractor to the purchases component, because of the different types between assigned employee and normal cards. Could definitely do that better.
also fixes the UI of the employee list too to display vertically
Unassignment button itself still clips out, but it's better, and that button is just a quick way to test it