docs: make end step position consistent
style button to be blue
make factory a bit nicer
still needs a colour schema, but I think I have all I need to get started on developign the components
There isn't much of a grid structure either on the top drawer - message takes up too much space and it's a bit inconsistent. Unfortunately, Invision is incredibly fragile, and it makes it very difficult to work with in small batches.
So, I'll export to images, set it up the hot spots manually and go from there.