refactor: use vuex for state management
This will become important for inter-frame communication, because passing events via frames just doesn't seem the right way to do it.
Pretty hacky and quick implementation of the Vuex, but it's migration is done in a minimal way to help preserve the existing code.
After implementing the Banner frame, I can refactor this.