feat: hydrate the card with real data
refactored the layout for simpler and better arrows
renamed cotton/linen to be consistent with underlying data type (wool)
ensured placeholder image was correctly referencing the placeholder type (not unknown)
Add aggregation counts of the resources - resources are currently just flat items, and have no means of aggregating them.
Simpler in some ways for resource management, but not simpler here! Might need to refactor this to expose this aggregation mechanic for other parts of the code

secondary resources aren't displayed here yet.