feat: add market banner
deprecate the old market card
Create a new reusable component to display a collection of primary resources
add a default ordering of those primary resources
add flags for primary resource collection component. Most of this isn't needed, but could be nice to have in the future. And it will related to the secondary resource collection too.