feat: add purchasing banner
includes a placeholder employee component (no interactibility, but the elements are there)
Couldn't name it the same because of naming conflicts with the existing component
the flexbox styling of the separator was a faff - beware refactoring this. v-template doesn't work with flexbox, and ListView imposes its own styling. Must be a way around it, but I didn't look too much into it.