feat: add rad side drawers
just placeholders for the left/right sides
RadSideDrawer not good at handling multiple bars, so there was a bit of restructuring needed for its directives to work
and there are manual hooks into the swipe events. A bit of a compromise. There are packages out there that look like they handle it a little better. But not much.

Also, the grass border was lost in this process. It's likely this will need to be an embedded image to the banner, or included directly in the banner component, unfortunately.