chore: migrate to NS7
Uses ns migrate, as well as several steps to fix the CSS
The tests are still broken in the webpack build, despite everything saing they should be excluded.
Shadow label couldn't be carried across, it's not compatible
Also the full screen is no longer working.
Not a trouble free migration!