feat: display custom components for game messages
filter down the messages according to the log level - this excludes sent messages, and limits it to only the interesting items.
delete the old event icon/message components
use more appropriate game icons.
Move the variable stylings into global css - hopefully this and the other changes helps with the performance issue I was seeing.
Limit the messages to 6 of the most recent, and no longer make them scrollable.
it would be nice to be able to aggregate the messages, but this is probably good enough now for the first iteration